requirements.txt
lightning
transformers
diffusers==0.30.3
hf_transfer
sentencepiece
protobuf
schedulefree
accelerate
peft
optimum-quanto
bitsandbytes
prodigyopt
huggingface_hub==0.22.2
Pillow
pathlib
wandb
mistral_inference
mistral_common